% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/as.data.table.tbl.R
\name{as.data.table.tbl}
\alias{as.data.table.tbl}
\title{Convert a dplyr table to a data.table by stripping the dplyr classes}
\usage{
\method{as.data.table}{tbl}(x)
}
\arguments{
\item{x}{dplyr tbl object}
}
\description{
Convert a dplyr table to a data.table by stripping the dplyr classes
}

