% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autokey.R
\name{autokey}
\alias{autokey}
\alias{set_autokeys}
\alias{autokeys}
\title{Automatically set data.table keys}
\usage{
autokey(x, keys = getOption("autokey"))

set_autokeys(cols)

autokeys()
}
\arguments{
\item{x}{data.table that will get autokeys set.}

\item{keys}{character; the values to set for keys if present in \code{dt}.

\code{autokey} sets the keys for a data.table. The default is to look at the
\code{getOption('autokey')} for the list available keys. It is equivalent
to setting the \code{cols} argument of \code{setkeyv} equivalent to:

\code{ intersect( names(x), getOption('autokey') ) }}

\item{cols}{character; column names to be automatically set as keys

This is nothing more than a wrapper for options(autokey=x)}
}
\description{
Atomatically sets the keys of a data.table based on \code{options}.

Set the column for autokeys

Get the vector of autokeys
}
\details{
Simple wrapper of \code{getOption('autokey')}
}
\examples{
  options( autokey='Species' )
  data(iris)
  setDT(iris)
  autokey(iris)

}
\seealso{
\code{setkeyv} in the \code{data.table} package.
}
