% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_list.R
\name{by_list}
\alias{by_list}
\title{Turn a list of variables into a by list}
\usage{
by_list(.data, x)
}
\value{
names list; values are the names
}
\description{
Turn a list of variables into a by list
}
\examples{

mpg \%>\% by_list( c("year","trans" ) )

}
