% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropicols.R
\name{dropicols}
\alias{dropicols}
\title{dropicols}
\usage{
dropicols(x)
}
\arguments{
\item{x}{data.table}
}
\description{
Drop columns prefixed by \verb{i.*} these may be remenants of a join.
}
\details{
During a merge/join matching columns in the two merge tables get a prefix.
The i-argument recieves a prefix \code{i.}; these colmns may not be wanted.
\code{dropicols} removes these columns from a data table.
}
