% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_left_join.r
\name{dt_left_join}
\alias{dt_left_join}
\title{dt_left_join}
\usage{
dt_left_join(x, y, by, ...)
}
\arguments{
\item{x, y}{data.tables to join}

\item{by}{chaacter; colums to match can by named}
}
\value{
a data.tale
}
\description{
Perform a dt
}
\details{
columns are reorder so that
}
\examples{

lhs <- data.table( letters=letters[1:10], i=1:10 )
rhs <- data.table( letters=letters[1:5], x=5:1 )

# setkey(lhs, letters )
# setkey(rhs, letters )

lhs \%>\% dt_left_join( rhs, by="letters" )
lhs \%>\% dt_left_join( rhs, by="letters" )

}
\seealso{
\code{\link[data.table]{data.table}}
}
