% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_table.R
\name{i_table}
\alias{i_table}
\title{Create a table for use in data.tables i slot}
\usage{
i_table(.data, x)
}
\arguments{
\item{.data}{data.table}

\item{x}{character; name of character columns}
}
\description{
Create a table for use in data.tables i slot
}
