% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keysfirst.R
\name{keysfirst}
\alias{keysfirst}
\title{keysfirst}
\usage{
keysfirst(data)
}
\arguments{
\item{data}{data to return}
}
\value{
character vector; the names
}
\description{
data.table names with the keys expressed fist
}
\examples{
  \dontrun{
    dt \%>\% setcolorder( keysfirst(dt))
  }


}
\seealso{
\code{\link[base]{names}}
  \code{\link[data.table]{key}}
}
