% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join.R
\name{[.data.tableoff}
\alias{[.data.tableoff}
\title{left_join}
\usage{
\method{[}{data.tableoff}(x, i, ...)
}
\arguments{
\item{x}{data.table}

\item{i}{object}
}
\description{
implements a default left join for data.table
}
\details{
By convention DT implements \code{DT[X]} as a RIGHT JOIN. This convention promotes
facts, i.e. the most quickly varying values to be on the RHS of the result.
This is anathema to the convention of modeling where the Y's are expressed
first.
`
}
\examples{

 x <- data.table( letters=letters, i=1:26 )
 setkey(x, letters )

 rhs <- data.table( letters=letters[1:5], i=1:5 )
 setkey(rhs, letters )

 data.table:::`[.data.table`(x,rhs)      # RIGHT JOIN

 options( datatable.default.left.join = FALSE )
 x[rhs, on=.(letters)]                     # RIGHT JOIN
 options( datatable.default.left.join = TRUE )
 x[rhs, on=.(letters)]

}
