% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge0.R
\name{merge0}
\alias{merge0}
\title{Left (outer) join of data.tables replacing introduced NAs with default values}
\usage{
merge0(x, y, default = 0, ...)
}
\arguments{
\item{x}{left data.table}

\item{y}{right data.table}

\item{default}{value used where records between tables do not match}

\item{...}{additional arguments passed to other functions}
}
\description{
Implements a left join between data.tables setting default values for
variables where the join is unequal.
}
\examples{
  \dontrun{
    data(iris)
    setDT(iris)
    merge0( iris, iris[,-3])
  }
}
\seealso{
`left_join`
}
