% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_keys_first.R
\name{names_keys_first}
\alias{names_keys_first}
\alias{keysfirst}
\title{names_keys_first}
\usage{
names_keys_first(data)

keysfirst(data)
}
\arguments{
\item{data}{data to return}
}
\value{
character vector; the names
}
\description{
data.table names with the keys expressed first
}
\examples{
  \dontrun{
    dt \%>\% setcolorder( keysfirst(dt) )
  }

}
\seealso{
\code{\link[base]{names}}
  \code{\link[data.table]{key}}
}
