% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageDT.R
\name{pageDT}
\alias{pageDT}
\title{page through a data.table by key}
\usage{
pageDT(x, by = key(x))
}
\arguments{
\item{x}{data.table}

\item{by}{character containing column names of x that should be paged through

pageDT allows the user to interacitvely page through a data.table stepping
through by the values of a key.

The commands generally follow those of \code{\link[base]{browser}}
n - next key
Q - quit}
}
\description{
page through a data.table by key
}
\note{
TODO
\itemize{
\item remove key columns
}
}
\examples{
  
  \dontrun{ 
 #  data(iris) 
 #   setDT( iris )
 #   setkey( iris, Species )
 #   pageDT( iris ) 
  }

}
\seealso{
\code{\link[base]{browser}}
\code{\link[base]{readline}}
}
\author{
Christopher Brown
}
