% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_replace.R
\name{prefix_replace}
\alias{prefix_replace}
\alias{prefix_replace.data.table}
\title{Replace Prefix}
\usage{
prefix_replace(x, prefix, replace)

\method{prefix_replace}{data.table}(x, prefix, replace)
}
\arguments{
\item{x}{object to have prefix replaced}

\item{prefix}{string or regex without the separator}

\item{replace}{string used as replacement}
}
\description{
Convenience function for replacing prefixes
}
\examples{

 dt <- data.table( i.a = 1:5, ii.b = 1:5)
 prefix_replace( dt, "i", "X." )

}
\seealso{
\itemize{
\item \code{dp.misc::prefix}
\item \code{base.tools::nm_prefix}
\item \code{stringr.tools::str_prefix}
}
}
