% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/print.data.table.R
\name{print.data.table}
\alias{print.data.table}
\title{print.data.table}
\usage{
\method{print}{data.table}(x, ...)
}
\arguments{
\item{x}{data.table}

\item{...}{arguments passed to}
}
\value{
NONE, used for side effects only
}
\description{
Extends \code{\link[data.table]{print}} by printing x followed by its key
}
\examples{
data(iris)
  setDT(iris)

  print(iris)

  setkey(iris, Species )
  print(iris)
}

