% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameif.R
\name{renameif}
\alias{renameif}
\title{renameif}
\usage{
renameif(x, old, new)
}
\arguments{
\item{x}{table}

\item{old}{string; old name}

\item{new}{string; new name@name}
}
\description{
Renames a column if already exists otherwise doesn't do a thing
}
\details{
Renames a column on the condition that it exists

Does not take mu
Made with pipes in mind.
}
\seealso{
\itemize{
\item \code{\link[dplyr:rename]{dplyr::rename()}}
}
}
