% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setkeysfirst.R
\name{setkeysfirst}
\alias{setkeysfirst}
\alias{setcolorderkeysfirst}
\title{setkeysfirst}
\usage{
setkeysfirst(x)

setcolorderkeysfirst(x)
}
\arguments{
\item{x}{data.table on which to reorder columns}
}
\value{
character vector with columns reodrdered so the key columns are first.
}
\description{
Reorder columns to put the key columns first.
}
\details{

}
\examples{

 data(iris)
 setDT(iris)
 setkey(iris, Species)
 setkeysfirst(iris)
 iris

}
\seealso{
\code{\link[data.table]{key}},
\code{\link[data.table]{setcolorder}}
}
