% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setprefix.R
\name{setprefix}
\alias{setprefix}
\title{setprefix}
\usage{
setprefix(x, old, new)
}
\arguments{
\item{x}{data.table}

\item{old}{string possibly a regex (there is no need to add "^")}

\item{new}{string}
}
\description{
set/change  prefix on column names
}
\details{
\code{setprefix} works similar to \code{\link[data.table]{setnames}} but
works only on the beginning of names not the whole names by swapping out the
beginning.
}
\examples{
x <-data.table( i.x=1:10, i.y=1:10 )
x \%>\% setprefix( "i.", "x." )
names(x)

}
\seealso{
\code{\link[data.table]{setnames}}
}
