% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.table-tools.R, R/tb.R
\name{tb}
\alias{tb}
\title{Print compact data.table::tables info}
\usage{
tb()

tb()
}
\description{
Prints data.tables with a good defaults

Shows compact data.table info
}
\note{
(c)2013 Decision Patterns
}
\seealso{
\code{\link[data.table]{tables}}
}

