% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.tbl.R
\name{as.data.table.tbl}
\alias{as.data.table.tbl}
\title{Convert a dplyr table to a data.table by stripping the dplyr classes}
\usage{
\method{as.data.table}{tbl}(x, keep.rownames = FALSE)
}
\arguments{
\item{x}{tbl object}

\item{keep.rownames}{logical; unused but kepy to match as.data.table generic}
}
\description{
Convert a dplyr table to a data.table by stripping the dplyr classes
}
\examples{

 tbl <- as.tibble(iris)
 dt <- as.data.table(iris)

 class(tbl)
 class(dt)

}
\seealso{
- [tibble::as.tibble()]
}
