% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entable.R
\name{entable}
\alias{entable}
\title{Convert vectors to data frames and vice versa}
\usage{
entable(x, name = ".key", value = as.character(substitute(x)))
}
\arguments{
\item{x}{atomic vector (named)}

\item{name, value}{Names of the columns that store the names and values.
If `name` is `NULL`, a one-column tibble is returned; `value` cannot be `NULL`.}
}
\value{
A [data.table] with two columns (if `name` is not `NULL`, the default)
  or one column (otherwise).#'
}
\description{
Convert vectors to data frames and vice versa
}
\examples{
  x = 1:3
  entable(x)
  names(x) <- letters[x]
  entable(x)
  entable(x, name="letter")

  entable(x)

}
\seealso{
`tibble::enframe()`
}
