% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonkeys.R
\name{nonkey}
\alias{nonkey}
\alias{nonkeys}
\title{Get non-keys from a table}
\usage{
nonkey(x)

nonkeys(x)
}
\arguments{
\item{x}{a data set that may have a key}
}
\value{
character
}
\description{
Return the names of non-key columns
}
\examples{
data(iris)
setDT(iris)

nonkeys(iris)

setkey(iris, Species )
key(iris)
nonkey(iris)
setkey( iris, Species, Sepal.Length )

}
