% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.data.table.R
\name{print.data.table}
\alias{print.data.table}
\title{print.data.table}
\usage{
\method{print}{data.table}(x, topn = getOption("datatable.print.topn"),
  nrows = getOption("datatable.print.nrows"),
  class = getOption("datatable.print.class"),
  row.names = getOption("datatable.print.rownames"),
  col.names = getOption("datatable.print.colnames"),
  print.keys = getOption("datatable.print.keys"), quote = FALSE, ...)
}
\arguments{
\item{x}{data.table}

\item{...}{arguments passed to}
}
\value{
NONE, used for side effects only
}
\description{
Extends \code{print.data.table} by printing x followed by its key.
}
\details{
This is just a copy of [data.table:::print.data.table()] that uses a message
[keys_msg()] to display the keys when printed.
}
\examples{
  data(iris)
  setDT(iris)

  print(iris)

  setkey(iris, Species )
  print(iris)

  iris[ , rn := 1 ]


}
\references{
[data.table:::print.data.table()]
}
