% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setcolsfirst.R
\name{setcolfirst}
\alias{setcolfirst}
\title{Move columns to the beginning}
\usage{
setcolfirst(x, firstcols)
}
\arguments{
\item{firstcols}{vector; character (names of columns) integer
  (column numbers)}

\item{a}{data.table}
}
\description{
Move columns to the beginning
}
\examples{
data(iris)
  setDT(iris)

  setcolfirst(iris,"Species")
}

